Cars making proper turns
The "gears" have not been merged into this yet, but I figured this part
out and wanted to share it before continuing on to doing the other bits

For a stress test I have the spawner set to spawn many many cars, and
already we're seeing stuff like congestion emerging (cars driving slowly
near intersections)

The TrafficAgent is alas inactive right now. The code assumes lights for
each possible turn and I didn' want to mess with it but it's still
there.