Better turning logic
Turn points are now either the intersection point of the two lanes (as
before) OR the midway point between the end of one lane and the start of
the next. The second happens if the lanes are parallel, or the
intersection point is within either lane.

Cars
- Car now keep track of their own direction of travel
- CarAI handles that cars drive from lane end to turn point, and then
from turn point to start of next lane
- NOTE: distanceOnLane might be slightly inaccurate now